+++ Assets/Tests/Behavior/ContainerBuilderOverrideTests.cs
@@ -11,7 +11,7 @@ namespace MrWatts.Internal.FuelInject.TestProject.Tests.Behaviour
     internal sealed class ContainerBuilderOverrideTests : SceneTestGroup
     {
         [UnityTest]
-        public IEnumerator ContainerBindingsCanBeOverriddenDuringTests()
+        public IEnumerator ContainerBindingsCanBeOverriddenDuringTestsWithAutomaticallyAddRootGameObjectModulesEnabled()
         {
             bool isOverrideInvoked = false;
 
@@ -35,6 +35,49 @@ namespace MrWatts.Internal.FuelInject.TestProject.Tests.Behaviour
                 "Container module override handlers do not properly override existing bindings. Perhaps they " +
                 "aren't loaded after the actual scene modules and are thus overridden again by them?"
             );
+        }
+
+        [UnityTest]
+        public IEnumerator ContainerBindingsCanBeOverriddenDuringTestsWithAutomaticallyAddRootGameObjectModulesDisabled()
+        {
+            bool isOverrideInvoked = false;
+
+            yield return SetupScene(
+                "TestSceneExplicitModules",
+                builder =>
+                {
+                    isOverrideInvoked = true;
+
+                    builder.RegisterType<CustomBar>().As<IBar>().SingleInstance();
+                }
+            );
+
+            Assert.IsTrue(isOverrideInvoked, "Registered container module override handlers are not invoked");
+
+            var service = GameObjectFinder.Get<MonoBehaviourWithBarDependency>();
+
+            Assert.IsInstanceOf(
+                typeof(CustomBar),
+                service.BarGetter,
+                "Container module override handlers do not properly override existing bindings. Perhaps they " +
+                "aren't loaded after the actual scene modules and are thus overridden again by them?"
+            );
+        }
+
+        [UnityTest]
+        public IEnumerator GlobalOverridingModuleIsOnlyRegisteredOnceWithAutomaticallyAddRootGameObjectModulesEnabled()
+        {
+            int timesOverrideInvoked = 0;
+
+            yield return SetupScene(
+                "TestScene",
+                _ =>
+                {
+                    ++timesOverrideInvoked;
+                }
+            );
+
+            Assert.AreEqual(1, timesOverrideInvoked, "Global overriding modules should only be registered once");
         }
     }
 }